queries

+++ contracts/whitelist-updatable/src/state.rs
@@ -11,4 +11,4 @@ pub struct Config {
 pub const CONFIG: Item<Config> = Item::new("config");
 pub const TOTAL_ADDRESS_COUNT: Item<u64> = Item::new("total_address_count");
 // Holds all addresses and mint count
-pub const WHITELIST: Map<Addr, u64> = Map::new("wl");
+pub const WHITELIST: Map<Addr, u32> = Map::new("wl");
